Extract readLine function from Parser
