+++ src/main/kotlin/ru/byprogminer/shellin/Parser.kt
@@ -1,10 +1,9 @@
 package ru.byprogminer.shellin
 
 import ru.byprogminer.shellin.command.*
+import ru.byprogminer.shellin.util.readLine
 import java.io.BufferedInputStream
 import java.io.EOFException
-import java.nio.ByteBuffer
-import java.nio.channels.Channels
 
 
 /**
@@ -15,8 +14,6 @@ class Parser(
 ) {
 
     private companion object {
-
-        const val READ_LINE_BUFFER_SIZE = 1024
 
         val VARIABLE_REGEX = "\\\$([a-zA-Z_]\\w*)".toRegex()
         val ASSIGNMENT_REGEX = "([a-zA-Z_]\\w*)=(.*)".toRegex()
@@ -43,7 +40,7 @@ class Parser(
      * @throws IllegalArgumentException if first line is malformed command
      */
     fun parse(input: BufferedInputStream): Command? {
-        val line = readLine(input) ?: throw EOFException()
+        val line = input.readLine() ?: throw EOFException()
 
         val chunks = prepareLine(line)
         require(chunks.isNotEmpty())
@@ -205,77 +202,5 @@ class Parser(
      */
     private fun substituteVariables(str: String) = VARIABLE_REGEX.replace(str) { match ->
         state.environment.getOrDefault(match.groups[1]!!.value, "")
-    }
-
-    /**
-     * Reads line from input stream with preserving trailing bytes.
-     *
-     * If EOF reached returns `null`.
-     * End of line is LF. Encoding is UTF-8.
-     * EOL is preserved in the end of returned string.
-     *
-     * @param input input stream to read from
-     * @return read line or `null` if EOF
-     */
-    private fun readLine(input: BufferedInputStream): String? {
-        require(input.markSupported())
-
-        val channel = Channels.newChannel(input)
-        var buffer = ByteBuffer.allocate(READ_LINE_BUFFER_SIZE)
-
-        while (true) {
-            val pos = buffer.position()
-
-            // assert invariant
-            require(buffer.hasRemaining())
-
-            // mark current position to reset to when line will read
-            input.mark(buffer.remaining())
-
-            // read up to buffer remaining
-            val wasRead = channel.read(buffer)
-
-            // if EOF - there isn't any line, return null
-            // this isn't exactly -1 to prevent infinite loop
-            if (wasRead <= 0) {
-                return null
-            }
-
-            // try to find LF in read bytes
-
-            var ok = false
-            for (i in pos until buffer.position()) {
-                if (buffer[i] == '\n'.code.toByte()) {
-                    // if found - break the loop and save position
-
-                    ok = true
-                    buffer.flip()
-                    buffer.position(i + 1)
-                    break
-                }
-            }
-
-            // if LF found reset input to mark and skip bytes to EOL
-            if (ok) {
-                input.reset()
-                input.skip((wasRead - buffer.remaining()).toLong())
-                break
-            }
-
-            // if we reached end of buffer, reallocate with multiplication strategy
-            // to get amortized complexity O(n)
-            if (!buffer.hasRemaining()) {
-                buffer.flip()
-
-                val newBuffer = ByteBuffer.allocate(buffer.capacity() * 2)
-                newBuffer.put(buffer)
-                buffer = newBuffer
-            }
-        }
-
-        // decode read line with UTF-8
-
-        buffer.flip()
-        return Charsets.UTF_8.decode(buffer).toString()
     }
 }
